Move endpoint-specific entities in API files
As mostly a legacy decision, all entities are in one file.
But they're better off where endpoint-specific entities
can hang out in the entity file.
